Version 1.0

+++ Nomina/src/main/java/IPagable.java
@@ -7,5 +7,5 @@ public interface IPagable {
 
     double calcularSueldo(double bono);
 
-    double calcularSueldo(double bono, double horasExtra);
+    double calcularSueldo(double bono, int horasExtra);
 }
